requirements.txt
streamlit>=1.30.0
Pillow
